choeuna/choeuna1/requirements.txt
pymongo
flask
requests
python-dotenv